Wishlistdetail update delete

+++ Backend/WebshopApi/Controllers/WishlistController.cs
@@ -26,25 +26,25 @@ public class WishlistController : ControllerBase
     [HttpGet]
     public async Task<ActionResult<List<Wishlists>>> GetAllWishlists()
     {
-        var wishlists = _wishlistservice.GetAllWishLists();
+        var wishlists = await _wishlistservice.GetAllWishLists();
         return Ok(wishlists);
     }
 
     [HttpGet("{id:int}")]
-    public async Task<ActionResult<Wishlists>> GetWishlistById(int id)
+    public async Task<ActionResult<List<Wishlists>>> GetWishlistById(int id)
     {
-        var wishlist = _wishlistservice.GetWishlistsById(id);
+        var wishlist = await _wishlistservice.GetWishlistsById(id);
         return Ok(wishlist);
     }
     [Authorize]
     [HttpGet("mine")]
-    public async Task<ActionResult<Wishlists>> GetWishlistById()
+    public async Task<ActionResult<Wishlists>> GetMyWishlist()
     {
         var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
         // return Ok(useridstring);
         if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
         var userId = int.Parse(useridstring);
-        var wishlists = await _wishlistservice.GetWishlistsById(userId);
+        var wishlists = await _wishlistservice.GetWishlistsByUserId(userId);
         return Ok(wishlists);
     }
     [Authorize]
@@ -56,21 +56,37 @@ public class WishlistController : ControllerBase
         if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
         int userid = int.Parse(useridstring);
         wishlistDTO.UserId = userid;
-        _wishlistservice.CreateService(wishlistDTO);
-        return Ok();
+        var result = await _wishlistservice.CreateService(wishlistDTO);
+        return Ok(result);
     }
 
     [HttpPut("update")]
     public async Task<ActionResult> UpdateWishlist(WishlistDTO wishlistDTO)
     {
-        _wishlistservice.UpdateService(wishlistDTO);
+        await _wishlistservice.UpdateService(wishlistDTO);
         return Ok();
     }
 
-    [HttpDelete("delete")]
-    public async Task<ActionResult> DeleteWishlist([FromBody] int id)
+    [Authorize]
+    [HttpDelete("delete/{id:int}")]
+    public async Task<ActionResult> DeleteWishlist([FromRoute] int id)
     {
-        _wishlistservice.DeleteService(id);
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        int userid = int.Parse(useridstring);
+        await _wishlistservice.DeleteService(id);
+        return Ok();
+    }
+    [Authorize]
+    [HttpDelete("delete/product")]
+    public async Task<ActionResult> DeleteProductFromWishlist([FromBody] DeleteProductDto deleteProductDto)
+    {
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        int userid = int.Parse(useridstring);
+        await _wishlistservice.DeleteProduct(deleteProductDto.Id, deleteProductDto.WishlistName);
         return Ok();
     }
 }
